+++ include/Game/System/GameDataHolder.hpp
@@ -28,7 +28,7 @@ public:
     bool isOnGameEventValueForBit(const char*, int) const;
     void setGameEventValueForBit(const char*, int, bool);
     s32 getPictureBookChapterCanRead() const;
-    u32 getPictureBookChapterAlreadyRead() const;
+    u16 getPictureBookChapterAlreadyRead() const;
     void setPictureBookChapterAlreadyRead(int);
     void setRaceBestTime(const char*, u32);
     u32 getRaceBestTime(const char*) const;
@@ -55,7 +55,7 @@ public:
     s32 setupSpinDriverPathStorage(const char*, int, int, int, f32*);
     void updateSpinDriverPathStorage(const char*, int, int, f32);
     u16 getStarPieceNumGivingToTicoSeed(int) const;
-    u32 getStarPieceNumMaxGivingToTicoSeed(int) const;
+    u16 getStarPieceNumMaxGivingToTicoSeed(int) const;
     void addStarPieceGivingToTicoSeed(int, int);
     bool isCompleteMarioAndLuigi() const;
     bool isPassedStoryEvent(const char*) const;
